+++ easy-fs/src/vfs.rs
@@ -57,7 +57,7 @@ impl Inode {
                 DIRENT_SZ,
             );
             if dirent.name() == name {
-                return Some(dirent.inode_number() as u32);
+                return Some(dirent.inode_number());
             }
         }
         None
